refactor: switch to the signed entity type messages
- remove previous `messages_parts/signed_entity_type.rs`
- rename `new` file with previous name
- adapt code
