+++ mithril-common/src/messages/message_parts/mod.rs
@@ -2,7 +2,6 @@ mod cardano_transactions_set_proof;
 mod certificate_metadata;
 mod mk_set_proof;
 mod signed_entity_type;
-mod signed_entity_type_new;
 mod signer;
 
 pub use cardano_transactions_set_proof::CardanoTransactionsSetProofMessagePart;
