+++ mithril-test-lab/mithril-end-to-end/src/stress_test/payload_builder.rs
@@ -9,7 +9,8 @@ use mithril_cardano_node_internal_database::{
 use mithril_common::{
     StdResult,
     entities::{
-        CardanoDbBeacon, Epoch, ProtocolMessage, ProtocolMessagePartKey, ProtocolParameters, Signer,
+        CardanoDbBeacon, Epoch, ProtocolMessage, ProtocolMessagePartKey, ProtocolParameters,
+        SignedEntityType, Signer,
     },
     messages::{RegisterSignatureMessageHttp, RegisterSignerMessage, SignedEntityTypeMessage},
     protocol::ToMessage,
@@ -93,8 +94,8 @@ pub async fn compute_mithril_stake_distribution_signatures(
                     .sign_all(&mithril_stake_distribution_message)
                     .into_iter()
                     .map(|s| RegisterSignatureMessageHttp {
-                        signed_entity_type: SignedEntityTypeMessage::MithrilStakeDistribution(
-                            epoch,
+                        signed_entity_type: SignedEntityTypeMessage::Known(
+                            SignedEntityType::MithrilStakeDistribution(epoch),
                         ),
                         party_id: s.party_id.clone(),
                         signature: s.signature.clone().to_json_hex().unwrap(),
@@ -165,8 +166,11 @@ pub async fn compute_immutable_files_signatures(
                     .sign_all(&cardano_immutable_files_full_message)
                     .into_iter()
                     .map(|s| RegisterSignatureMessageHttp {
-                        signed_entity_type: SignedEntityTypeMessage::CardanoDatabase(
-                            CardanoDbBeacon::new(*epoch, immutable_file_number),
+                        signed_entity_type: SignedEntityTypeMessage::Known(
+                            SignedEntityType::CardanoDatabase(CardanoDbBeacon::new(
+                                *epoch,
+                                immutable_file_number,
+                            )),
                         ),
                         party_id: s.party_id.clone(),
                         signature: s.signature.clone().to_json_hex().unwrap(),
